Split LocatorSerializableObjectFilterIntegrationTest for Java 8 and 9
(cherry picked from commit 2b31394f88ef5e7bdf9a6865621b3087594fcc93)
